#3770 move LocalePicker to a separate component
