+++ src/pages/settings/PreferencesPage.js
@@ -4,6 +4,7 @@ import {withOnyx} from 'react-native-onyx';
 import PropTypes from 'prop-types';
 
 import HeaderWithCloseButton from '../../components/HeaderWithCloseButton';
+import LocalePicker from '../../components/LocalePicker';
 import Navigation from '../../libs/Navigation/Navigation';
 import ROUTES from '../../ROUTES';
 import ONYXKEYS from '../../ONYXKEYS';
@@ -12,7 +13,6 @@ import Text from '../../components/Text';
 import NameValuePair from '../../libs/actions/NameValuePair';
 import CONST from '../../CONST';
 import {setExpensifyNewsStatus} from '../../libs/actions/User';
-import {setLocale} from '../../libs/actions/App';
 import ScreenWrapper from '../../components/ScreenWrapper';
 import Switch from '../../components/Switch';
 import Picker from '../../components/Picker';
@@ -29,20 +29,16 @@ const propTypes = {
         expensifyNewsStatus: PropTypes.bool,
     }),
 
-    /** Indicates which locale the user currently has selected */
-    preferredLocale: PropTypes.string,
-
     ...withLocalizePropTypes,
 };
 
 const defaultProps = {
     priorityMode: CONST.PRIORITY_MODE.DEFAULT,
     user: {},
-    preferredLocale: CONST.DEFAULT_LOCALE,
 };
 
 const PreferencesPage = ({
-    priorityMode, user, translate, preferredLocale,
+    priorityMode, user, translate,
 }) => {
     const priorityModes = {
         default: {
@@ -54,17 +50,6 @@ const PreferencesPage = ({
             value: CONST.PRIORITY_MODE.GSD,
             label: translate('preferencesPage.focus'),
             description: translate('preferencesPage.focusModeDescription'),
-        },
-    };
-
-    const localesToLanguages = {
-        default: {
-            value: 'en',
-            label: translate('preferencesPage.languages.english'),
-        },
-        es: {
-            value: 'es',
-            label: translate('preferencesPage.languages.spanish'),
         },
     };
 
@@ -112,17 +97,7 @@ const PreferencesPage = ({
                     <Text style={[styles.formLabel]} numberOfLines={1}>
                         {translate('preferencesPage.language')}
                     </Text>
-                    <View style={[styles.mb2]}>
-                        <Picker
-                            onChange={(locale) => {
-                                if (locale !== preferredLocale) {
-                                    setLocale(locale);
-                                }
-                            }}
-                            items={Object.values(localesToLanguages)}
-                            value={preferredLocale}
-                        />
-                    </View>
+                    <LocalePicker />
                 </View>
             </View>
         </ScreenWrapper>
@@ -141,9 +116,6 @@ export default compose(
         },
         user: {
             key: ONYXKEYS.USER,
-        },
-        preferredLocale: {
-            key: ONYXKEYS.PREFERRED_LOCALE,
         },
     }),
 )(PreferencesPage);
